extracts TaskExecutor tests in a separate class
