+++ src/Idephix/Console/InputFactory.php
@@ -13,10 +13,9 @@ class InputFactory
      */
     public function buildFromUserArgsForTask($arguments, Task $task)
     {
-        // $defaultArguments = array('command' => null);
+        $defaultArguments = array();
 
         foreach ($task->userDefinedParameters() as $parameter) {
-
             if ($parameter->isFlagOption()) {
                 $defaultArguments['--' . $parameter->name()] = $parameter->defaultValue();
             } else {
